requirements.txt
Package               Version
--------------------- -------
asgiref               3.2.10
astroid               2.4.2
colorama              0.4.3
Django                3.0.8
django-rest-framework 0.1.0
djangorestframework   3.11.0
isort                 4.3.21
lazy-object-proxy     1.4.3
mccabe                0.6.1
Pillow                7.2.0
pip                   20.1.1
pylint                2.5.3
pytz                  2020.1
setuptools            49.2.0
six                   1.15.0
sqlparse              0.3.1
toml                  0.10.1
typed-ast             1.4.1
wheel                 0.34.2
wrapt                 1.12.1
